refactor: [M3-7178] - Refactor Cypress test setup, cache API account request, fix hanging (#9757)
* Perform clean up for domain resources

* Fetch Linode account info at Cypress start, mock account request using cached info

* Refactor Cypress test and command set up

* Refactor OAuth App tests into a single spec

* Fix test hanging by disabling new Chrome headless implementation
