+++ packages/manager/cypress.config.ts
@@ -1,15 +1,16 @@
 /* eslint-disable no-console */
 import { defineConfig } from 'cypress';
 import { setupPlugins } from './cypress/support/plugins';
+import { configureBrowser } from './cypress/support/plugins/configure-browser';
 import { configureFileWatching } from './cypress/support/plugins/configure-file-watching';
 import { configureTestSuite } from './cypress/support/plugins/configure-test-suite';
-import { disableGoogleSafeBrowsing } from './cypress/support/plugins/disable-google-safe-browsing';
 import { discardPassedTestRecordings } from './cypress/support/plugins/discard-passed-test-recordings';
 import { loadEnvironmentConfig } from './cypress/support/plugins/load-env-config';
 import { nodeVersionCheck } from './cypress/support/plugins/node-version-check';
 import { regionOverrideCheck } from './cypress/support/plugins/region-override-check';
 import { vitePreprocess } from './cypress/support/plugins/vite-preprocessor';
 import { configureApi } from './cypress/support/plugins/configure-api';
+import { fetchAccount } from './cypress/support/plugins/fetch-account';
 import { fetchLinodeRegions } from './cypress/support/plugins/fetch-linode-regions';
 import { splitCypressRun } from './cypress/support/plugins/split-run';
 import { enableJunitReport } from './cypress/support/plugins/junit-report';
@@ -55,9 +56,10 @@ export default defineConfig({
         configureApi,
         configureFileWatching,
         configureTestSuite,
+        configureBrowser,
         vitePreprocess,
-        disableGoogleSafeBrowsing,
         discardPassedTestRecordings,
+        fetchAccount,
         fetchLinodeRegions,
         regionOverrideCheck,
         splitCypressRun,
